fix: remove legacy Alarm Modal component.
