+++ src/components/alarmAdmin/CreateAlarmModal/style.ts
@@ -1,11 +1,14 @@
+import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { colors } from '@sopt-makers/colors';
+import { fontsObject } from '@sopt-makers/fonts';
+import { IconArrowUpRight } from '@sopt-makers/icons';
 
 export const StAlarmModalWrapper = styled.section`
-  width: 50.4rem;
+  width: 64rem;
 
   & > main {
-    padding: 1.6rem 3rem 3.2rem 3rem;
+    padding: 2.6rem 3rem;
 
     & > .type_selector {
       display: flex;
@@ -140,4 +143,193 @@ export const StTextArea = styled.textarea`
   &:focus {
     cursor: default;
   }
+`;
+
+export const SelectWrapper = styled.div`
+  display: flex;
+  gap: 2rem;
+`;
+
+export const InputWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1.6rem;
+
+  margin-top: 1.6rem;
+`;
+
+export const SendTimeWrapper = styled.div`
+  display: flex;
+  gap: 2rem;
+
+  margin-top: 1.6rem;
+
+  & .react-datepicker-wrapper {
+    height: 0;
+  }
+
+  & .react-datepicker__navigation {
+    width: 3.2rem;
+  }
+`;
+
+export const AttachWrapper = styled.div`
+  width: 100%;
+  margin-top: 1.6rem;
+`;
+
+export const AvailableCheckLink = styled.a`
+  display: flex;
+  width: max-content;
+  align-items: center;
+
+  margin-top: 0.2rem;
+
+  border-bottom: 1px solid ${colors.gray100};
+  cursor: pointer;
+  color: ${colors.gray100};
+`;
+
+export const StyledIconArrowUpRight = styled(IconArrowUpRight)`
+  width: 1.6rem;
+`;
+
+export const AvailableCheckText = styled.span`
+  ${fontsObject.LABEL_5_11_SB}
+`;
+
+export const AttachOptionButtonList = styled.ul`
+  display: flex;
+  gap: 0.6rem;
+
+  margin-top: 0.8rem;
+`;
+
+export const OptionalInputWrapper = styled.div<{ attachOption: string }>`
+  width: 100%;
+  margin-top: ${({ attachOption }) => attachOption !== '첨부 안함' && `0.8rem`};
+`;
+
+export const triggerCSS = css`
+  background-color: ${colors.gray700};
+  width: 18rem;
+`;
+
+export const inputCSS = css`
+  & input {
+    background-color: ${colors.gray700};
+    width: 100%;
+  }
+`;
+
+export const textAreaCSS = css`
+  margin-top: 0.8rem;
+  & div {
+    background-color: ${colors.gray700};
+    width: 100%;
+    & textarea {
+      background-color: ${colors.gray700};
+    }
+  }
+`;
+
+export const sendSelectCSS = css`
+  width: 18rem;
+  margin-top: 0.8rem;
+  & button {
+    width: 100%;
+  }
+  & ul {
+    background-color: ${colors.gray600};
+    width: 100%;
+    & button {
+      width: 100%;
+    }
+    & button:hover {
+      background-color: ${colors.gray500};
+    }
+  }
+`;
+
+export const sendTriggerCSS = css`
+  background-color: ${colors.gray700};
+  width: 100%;
+`;
+
+export const reserveDateSelectCSS = css`
+  width: 28rem;
+  & button {
+    width: 100%;
+  }
+  & ul {
+    background-color: ${colors.gray600};
+    width: 100%;
+    & button {
+      width: 100%;
+    }
+    & button:hover {
+      background-color: ${colors.gray500};
+    }
+  }
+  & svg {
+  }
+`;
+
+export const datePickerWrapperCSS = css`
+  height: 4.8rem;
+`;
+
+export const reserveTimeSelectCSS = (isBannedTime: boolean) => css`
+  width: 28rem;
+  & button {
+    width: 100%;
+    & > div:not(ul div) {
+      border: ${isBannedTime && `1px solid ${colors.error}`};
+    }
+  }
+  & ul {
+    background-color: ${colors.gray600};
+    width: 100%;
+    & button {
+      width: 100%;
+    }
+    & button:hover {
+      background-color: ${colors.gray500};
+    }
+  }
+`;
+
+export const reserveTimeTriggerCSS = (isBannedTime: boolean) => css`
+  background-color: ${colors.gray700};
+  width: 100%;
+  & p {
+    color: ${isBannedTime && colors.error};
+  }
+`;
+
+export const reserveDateTriggerCSS = (selectedDate: Date | null) => css`
+  background-color: ${colors.gray700};
+  width: 100%;
+  & p {
+    color: ${selectedDate ? colors.white : colors.gray300};
+  }
+`;
+
+export const deepLinkSelectCSS = css`
+  width: 100%;
+  & button {
+    width: 100%;
+  }
+  & ul {
+    background-color: ${colors.gray600};
+    width: 100%;
+    & button:hover {
+      background-color: ${colors.gray500};
+    }
+  }
+`;
+
+export const deepLinkTriggerCSS = css`
+  background-color: ${colors.gray700};
+  width: 100%;
 `;
